Take webhook as 'with' instead of 'env' (#1)
It bothers me how this only takes one value as an `env` and everything else goes in as `with`. I get that the `env` is supposed to be environment variables... but I think making every value be passed into the GH action the same is distinctly easier and more consistent.